Moved example directory and code to root directory
