+++ example/src/LoginPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Login from '../Login.js';
+import Login from '../../Login.js';
 
 class LoginPage extends Component {
