+++ example/src/index.js
