+++ example/src/registerServiceWorker.js
